pyproject.toml
[project]
name = "resume-parser"
version = "0.1.0"
description = "A professional-grade resume parser using multi-strategy extraction (Regex, NER, LLM)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Brijesh Patel", email = "brijkpatel@example.com" }]
keywords = ["resume", "parser", "nlp", "ner", "llm", "extraction"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "gliner==0.2.22",
    "google-generativeai==0.8.5",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "pdfminer.six==20250506",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
# Python path configuration - makes tests work regardless of environment
pythonpath = ["src"]

# Test discovery patterns
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Test output options
testpaths = ["src"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]

# Markers for organising tests
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests (uses real models, slow)",
    "slow: Slow running tests",
]

[dependency-groups]
dev = [
    "pytest==8.4.2",
    "reportlab>=4.4.10",
]
